Rename componentDefs to overrides
