+++ libs/wrapper-gen/src/generateWebTypes.ts
@@ -1,4 +1,4 @@
 import ComponentRegister from './generator/ComponentRegister';
-import '../componentsDef';
+import '../componentOverrides';
 
 ComponentRegister.generateWebTypes();
